Do not set `test_timeout' in `unix_load'
The `test_timeout' global parameter is expected to be optionally set by
a board description file and then taken into account by `unix_load' to
override the default of 300.  However not only the procedure checks for
the override but it sets the global parameter to the default as well if
not present.  This might make a minuscule TCL interpreter execution time
reduction, but seems rather awkward from the programming style's point
of view, and may cause undesired effects if the parameter is referred
elsewhere.

Remove the setting of the global parameter then, while retaining its
semantics.

	* config/unix.exp (unix_load): Don't ever set `test_timeout',
	just use it locally.