Update most license notices to refer to WWW address
The old GPL notice previously used in DejaGnu contained instructions to
write to the FSF to request a copy of the GPL.  Unfortunately, the FSF
office has moved, and the FSF now recommends citing a URL as a more
stable address from which users can obtain a copy of the GPL if needed.

This patch was automated.